Move QML models into sub package
