+++ geometry_constructor/application.py
@@ -6,13 +6,13 @@ Loading this module also registers with QML the required custom classes to load 
 
 import sys
 from os import path
-from geometry_constructor.instrument_model import InstrumentModel, SingleComponentModel
-from geometry_constructor.geometry_models import CylinderModel, OFFModel
 from geometry_constructor.json_loader import JsonLoader
 from geometry_constructor.json_writer import JsonWriter
-from geometry_constructor.pixel_models import PixelGridModel, PixelMappingModel, SinglePixelModel
-from geometry_constructor.qml_json_model import FilteredJsonModel
-from geometry_constructor.transform_model import TransformationModel
+from geometry_constructor.qml_models.geometry_models import CylinderModel, OFFModel
+from geometry_constructor.qml_models.instrument_model import InstrumentModel, SingleComponentModel
+from geometry_constructor.qml_models.json_model import FilteredJsonModel
+from geometry_constructor.qml_models.pixel_models import PixelGridModel, PixelMappingModel, SinglePixelModel
+from geometry_constructor.qml_models.transform_model import TransformationModel
 from geometry_constructor.validators import NameValidator, TransformParentValidator, NullableIntValidator
 from geometry_constructor.writers import HdfWriter, Logger
 from PySide2.QtCore import QUrl, QObject
