+++ geometry_constructor/qml_models/geometry_models.py
@@ -1,5 +1,5 @@
 """
-ListModel implementations for accessing and manipulating Geometry models in QML
+ListModel implementations for accessing and manipulating Geometry qml_models in QML
 
 See http://doc.qt.io/qt-5/qabstractlistmodel.html#subclassing for guidance on how to develop these classes, including
 what signals need to be emitted when changes to the data are made.
@@ -7,10 +7,10 @@ what signals need to be emitted when changes to the data are made.
 
 from geometry_constructor.data_model import CylindricalGeometry, OFFGeometry
 from geometry_constructor.geometry_loader import load_geometry
-from geometry_constructor.models import change_value
+from geometry_constructor.qml_models import change_value
 from PySide2.QtCore import Qt, QAbstractListModel, QModelIndex, QUrl, Signal, Slot
 
-from geometry_constructor.instrument_model import InstrumentModel
+from geometry_constructor.qml_models.instrument_model import InstrumentModel
 
 
 class CylinderModel(QAbstractListModel):
