+++ geometry_constructor/qml_models/pixel_models.py
@@ -4,10 +4,10 @@ ListModel implementations for accessing and manipulating pixel data in QML
 See http://doc.qt.io/qt-5/qabstractlistmodel.html#subclassing for guidance on how to develop these classes, including
 what signals need to be emitted when changes to the data are made.
 """
-from geometry_constructor.geometry_models import OFFModel
+from geometry_constructor.qml_models.geometry_models import OFFModel
 from geometry_constructor.data_model import PixelMapping, PixelGrid, SinglePixelId, Corner, CountDirection
-from geometry_constructor.instrument_model import InstrumentModel
-from geometry_constructor.models import change_value
+from geometry_constructor.qml_models.instrument_model import InstrumentModel
+from geometry_constructor.qml_models import change_value
 from PySide2.QtCore import Qt, QAbstractListModel, QModelIndex, Slot
 from math import isnan
